fix: KB mappings now cross-checked with Axe v3.4.0 rules (PR #299 Fixes #273)
