+++ packages/ace-report-axe/src/l10n/locales/en.json
@@ -11,6 +11,7 @@
         "aria-valid-attr": "ARIA",
         "aria-valid-attr-value": "ARIA",
         "button-name": "Forms",
+        "style": "Style",
         "checkboxgroup": "Forms",
         "color-contrast": "Color",
         "definition-list": "Lists",
@@ -51,7 +52,8 @@
         "th-has-data-cells": "Tables",
         "valid-lang": "Language",
         "video-caption": "Video",
-        "video-description": "Video"
+        "video-description": "Video",
+        "audio-caption": "Audio"
     },
     "nokb": "(no specific entry)"
 }
