+++ packages/ace-report-axe/src/l10n/locales/es.json
@@ -11,6 +11,7 @@
         "aria-valid-attr": "ARIA",
         "aria-valid-attr-value": "ARIA",
         "button-name": "Formularios",
+        "style": "Style",
         "checkboxgroup": "Formularios",
         "color-contrast": "Color",
         "definition-list": "Listas",
@@ -51,7 +52,8 @@
         "th-has-data-cells": "Tablas",
         "valid-lang": "Idioma",
         "video-caption": "Vídeo",
-        "video-description": "Vídeo"
+        "video-description": "Vídeo",
+        "audio-caption": "Audio"
     },
     "nokb": "(sin entrada específica)"
 }
